Updates to UI, cameras, and point system
Created a basic UI with a switch camera button and scoreboard

Added a third person camera which rotates around the bot when the mouse is dragged while the primary button is down. It also zooms with the scroll wheel. Found the script on Unity and made a few adjustments.

Changed coin to enable variable values for different kinds of coins.