Organized Files, Created Home Scene, Started Home Scene UI
The home scene will be the landing scene for the game. It will consist of a series of arena options. For each arena option, an input file can be selected from a dropdown menu. For now, pressing the start button will load the Arena scene.

The caveat here is that the user must have a folder on their desktop called "VEX_Simulator_Code". In that directory, the HomeUIController will look for .iqblocks files.

The next step will be to make the dropdown choices persistent in a JSON file.